Move keyboard avoidance to dynamic table view controller
