+++ src/xcode/ENA/ENA/Source/Scenes/DynamicTableViewController/DynamicTableViewController.swift
@@ -4,8 +4,10 @@
 
 import Foundation
 import UIKit
+import OpenCombine
 
 class DynamicTableViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
+	
 	var dynamicTableViewModel = DynamicTableViewModel([])
 
 	@IBOutlet private(set) lazy var tableView: UITableView! = self.view as? UITableView
@@ -44,7 +46,65 @@ class DynamicTableViewController: UIViewController, UITableViewDataSource, UITab
 		tableView.register(DynamicTableViewIconCell.self, forCellReuseIdentifier: DynamicCell.CellReuseIdentifier.icon.rawValue)
 		tableView.register(DynamicTableViewBulletPointCell.self, forCellReuseIdentifier: DynamicCell.CellReuseIdentifier.bulletPoint.rawValue)
 		tableView.register(DynamicTableViewHeadlineWithImageCell.self, forCellReuseIdentifier: DynamicCell.CellReuseIdentifier.headlineWithImage.rawValue)
+
+		setupKeyboardAvoidance()
 	}
+
+	private func setupKeyboardAvoidance() {
+		NotificationCenter.default.ocombine.publisher(for: UIApplication.keyboardWillShowNotification)
+			.append(NotificationCenter.default.ocombine.publisher(for: UIApplication.keyboardWillChangeFrameNotification))
+			.sink { [weak self] notification in
+
+				guard let self = self,
+					  let keyboardFrame = notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect,
+					  let animationDuration = notification.userInfo?[UIResponder.keyboardAnimationDurationUserInfoKey] as? Double,
+					  let animationCurveRawValue = notification.userInfo?[UIResponder.keyboardAnimationCurveUserInfoKey] as? Int,
+					  let animationCurve = UIView.AnimationCurve(rawValue: animationCurveRawValue) else {
+					return
+				}
+
+				var targetRect: CGRect?
+				if let currentResponder = self.view.firstResponder as? UIView {
+					let rect = currentResponder.convert(currentResponder.bounds, to: self.view)
+					if keyboardFrame.intersects(rect) {
+						targetRect = rect
+					}
+				}
+
+				let animator = UIViewPropertyAnimator(duration: animationDuration, curve: animationCurve) { [weak self] in
+					self?.tableView.scrollIndicatorInsets.bottom = keyboardFrame.height
+					self?.tableView.contentInset.bottom = keyboardFrame.height
+					if let targetRect = targetRect {
+						self?.tableView.scrollRectToVisible(targetRect, animated: false)
+					}
+				}
+				animator.startAnimation()
+			}
+			.store(in: &keyboardSubscriptions)
+
+		NotificationCenter.default.ocombine.publisher(for: UIApplication.keyboardWillHideNotification)
+			.sink { [weak self] notification in
+
+				guard let self = self,
+					  let animationDuration = notification.userInfo?[UIResponder.keyboardAnimationDurationUserInfoKey] as? Double,
+					  let animationCurveRawValue = notification.userInfo?[UIResponder.keyboardAnimationCurveUserInfoKey] as? Int,
+					  let animationCurve = UIView.AnimationCurve(rawValue: animationCurveRawValue) else {
+					return
+				}
+
+				let animator = UIViewPropertyAnimator(duration: animationDuration, curve: animationCurve) { [weak self] in
+					self?.tableView.scrollIndicatorInsets.bottom = 0
+					self?.tableView.contentInset.bottom = 0
+				}
+				animator.startAnimation()
+			}
+			.store(in: &keyboardSubscriptions)
+	}
+
+	// MARK: - Private
+
+	private var keyboardSubscriptions = Set<AnyCancellable>()
+
 }
 
 extension DynamicTableViewController {
